Refactor imports to centralize VSME_DEFAULTS from mireport.data.disclosures
