+++ tests/unitTests/xlsx_template_reader/conftest.py
@@ -3,8 +3,9 @@ from pathlib import Path
 import pytest
 
 from mireport.conversionresults import ConversionResultsBuilder
+from mireport.data.disclosures import VSME_DEFAULTS
 from mireport.taxonomy import loadBuiltInTaxonomyJSON
-from mireport.xlsx_template_reader.processor import VSME_DEFAULTS, XlsxProcessor
+from mireport.xlsx_template_reader.processor import XlsxProcessor
 from mireport.xlsx_template_reader.util import loadExcelFromPathOrFileLike
 
 SAMPLE = (
