+++ tests/unitTests/xlsx_template_reader/test_footnotes.py
@@ -10,7 +10,8 @@ from pathlib import Path
 import pytest
 
 from mireport.conversionresults import ConversionResultsBuilder
-from mireport.xlsx_template_reader.processor import VSME_DEFAULTS, XlsxProcessor
+from mireport.data.disclosures import VSME_DEFAULTS
+from mireport.xlsx_template_reader.processor import XlsxProcessor
 
 FOOTNOTE_SAMPLE = (
     Path(__file__).parents[3]
